AP starts! historical moment!
AP startup need trampoline code, which need to be placed under 1MB. But Wheel is located after 1MB, so we can use ld script to make this trampoline section relocate, and copy the code at runtime.

Multiprocessor startup is complicated, I did this by copying from a lot of resources.